refactor: move Skills.tsx to skills/skills.tsx and update import paths
