+++ src/Preview.tsx
@@ -5,7 +5,7 @@ import { InterestsPreview } from "./components/interests/interests-preview";
 import { PersonalDetailsPreview } from "./components/personal-details/personal-details-preview";
 import { QualificationsPreview } from "./components/qualifications/qualifications-preview";
 import { ReferencesPreview } from "./components/references/references-preview";
-import { SkillsPreview } from "./components/Skills";
+import { SkillsPreview } from "./components/skills/skills";
 
 export function Preview() {
   return (
