+++ src/components/skills/skills.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent } from "react";
 import { useForm } from "react-hook-form";
 import { create } from "zustand";
-import { Summary } from "./details/summary";
-import { Details } from "./details/details";
+import { Summary } from "../details/summary";
+import { Details } from "../details/details";
 
 type State = {
   skills: string;
